기존 auth 체크 로직이 적절하지 못해서, 수정
