--- src/constants/localstorage.ts
@@ -1,1 +1,0 @@
-export const ACCESS_TOKEN_KEY = "sb-dgsjfvfrgbrlfhpzedlo-auth-token";
