+++ src/utils/PublicRoute.tsx
@@ -1,4 +1,4 @@
-import { ACCESS_TOKEN_KEY } from "@/constants";
+import { useUserSession } from "@/hooks/useUserSession";
 import { LazyExoticComponent } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 
@@ -8,12 +8,13 @@ const PublicRoute = ({
   isAuthRequired?: boolean;
   LazyComponent?: LazyExoticComponent<() => JSX.Element>;
 }): React.ReactElement => {
-  const token = localStorage.getItem(ACCESS_TOKEN_KEY);
-  // 토큰이 있으면 회원가입페이지로 못가게 막는다.
+  const session = useUserSession();
+
   const { pathname } = useLocation();
   const navigate = useNavigate();
 
-  if (token && pathname === "/signup") {
+  // 로그인 후에는 회원가입페이지로 못가게 막는다.
+  if (session && pathname === "/sign-up") {
     navigate("/");
   }
 
@@ -21,4 +22,3 @@ const PublicRoute = ({
 };
 
 export default PublicRoute;
-// 여기까지 오지도 않는다.
