--- src/utils/getTokenFromLocalStorage.tsx
@@ -1,8 +1,0 @@
-import { ACCESS_TOKEN_KEY } from "@/constants";
-
-export const getTokenFromLocalStorage = () => {
-  const tokenWithUserJSON = localStorage.getItem(ACCESS_TOKEN_KEY);
-  if (!tokenWithUserJSON) return null;
-  const tokenWithUser = JSON.parse(tokenWithUserJSON);
-  return tokenWithUser.user.id;
-};
